shorten test names
